+++ app/drivers/api/endpoints/user.py
@@ -1,23 +1,12 @@
 from fastapi import Depends, HTTPException, APIRouter
 from sqlalchemy.orm import Session
-from typing import Optional
+from typing import Optional, List
 
 import app.domains.entities as entities
 import app.usecases as usecases
 from app.drivers.api.deps import get_user_usecase
-from app.drivers.rdb.models.test import TestUserTable
-from app.domains.entities.test import TestUser, TestUserCreate
 
 router = APIRouter()
-
-
-# @router.post("/", response_model=entities.User)
-# async def create_user(
-#     *,
-#     user_in: entities.UserCreate,
-#     uu: usecases.UserUsecase = Depends(get_user_usecase)
-# ) -> entities.User:
-#     return uu.create(obj_in=user_in)
 
 
 @router.post("/", response_model=entities.User)
@@ -52,7 +41,7 @@ async def update_user(
     return updated_user
 
 
-@router.delete("/", response_model=entities.User)
+@router.delete("/{id}", response_model=entities.User)
 async def delete_user(
     *, id: int, uu: usecases.UserUsecase = Depends(get_user_usecase)
 ) -> entities.User:
@@ -62,23 +51,19 @@ async def delete_user(
     return deleted_user
 
 
-@router.get("/", response_model=str)
-async def test_user() -> str:
-    return "aaa"
-
-
-# ユーザー情報取得(id指定)
-@router.get("/test/{id}")
-def get_user(user_id: int) -> TestUser:
-    user = Session.query(TestUserTable).filter(TestUserTable.id == user_id).first()
+@router.get("/{id}", response_model=entities.User)
+async def get_user(
+    *, id: int, uu: usecases.UserUsecase = Depends(get_user_usecase)
+) -> entities.User:
+    user: Optional[entities.User] = uu.read(id=id)
+    if user is None:
+        raise HTTPException(status_code=404)
     return user
 
+@router.get("/", response_model=entities.User)
+async def get_all_user(
+    *, uu: usecases.UserUsecase = Depends(get_user_usecase)
+) -> entities.User:
+    user: List[Optional[entities.User]] = uu.read_all()
+    return user
 
-# ユーザー情報登録
-@router.post("/test", response_model=TestUser)
-def post_user(user: TestUserCreate) -> TestUserCreate:
-    db_test_user = TestUserCreate(name=user.name, email=user.email)
-    Session.add(db_test_user)
-    Session.commit()
-    Session.refresh(db_test_user)
-    return db_test_user
